+++ packages/core/src/index.ts
@@ -28,7 +28,7 @@ export { BaseInstrumentation } from './instrumentations';
 export type { Meta, MetaGetter, Metas, MetaItem, App, User, Session } from './metas';
 
 export { getTransportBody, TransportItemType } from './transports';
-export type { Transport, TransportBody, TransportItem, TransportItemPayload, Transports } from './transports';
+export type { Transport, TransportBody, TransportItem, TransportItemPayload, Transports, BeforeSendHook } from './transports';
 export { BaseTransport } from './transports';
 
 export * from './utils';