+++ packages/core/src/transports/types.ts
@@ -37,4 +37,6 @@ export interface Transports {
   add: (...transports: Transport[]) => void;
   execute: (transportItem: TransportItem) => void;
   transports: Transport[];
+  pause: () => void;
+  unpause: () => void;
 }